web-service/requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==18.2.0
black==18.9b0
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SocketIO==3.2.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==5.0.0
peewee==3.8.2
pluggy==0.8.1
py==1.7.0
pytest==4.2.0
pytest-dotenv==0.3.1
python-dotenv==0.10.1
python-engineio==3.3.0
python-socketio==3.1.2
six==1.12.0
toml==0.10.0
Werkzeug==0.14.1
